+++ include/fullscore/transforms/clear_measure_transform.h
