+++ include/fullscore/transforms/erase_note_transform.h
