+++ include/fullscore/transforms/remove_dot_transform.h
