+++ src/actions/transforms/add_dot_transform_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/transforms/add_dot_transform_action.h>
 
-#include <fullscore/transforms/add_dot.h>
+#include <fullscore/transforms/add_dot_transform.h>
 #include <fullscore/models/note.h>
