+++ src/actions/transforms/clear_measure_transform_action.cpp
@@ -3,7 +3,7 @@
 
 #include <fullscore/actions/transforms/clear_measure_transform_action.h>
 
-#include <fullscore/transforms/clear_measure.h>
+#include <fullscore/transforms/clear_measure_transform.h>
 #include <fullscore/models/note.h>
