+++ src/actions/transforms/erase_note_action.cpp
@@ -4,8 +4,8 @@
 
 #include <fullscore/actions/transforms/erase_note_action.h>
 
-#include <fullscore/transforms/erase_note.h>
 #include <fullscore/models/note.h>
+#include <fullscore/transforms/erase_note_transform.h>
