+++ src/actions/transforms/insert_note_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/transforms/insert_note_action.h>
 
-#include <fullscore/transforms/insert_note.h>
+#include <fullscore/transforms/insert_note_transform.h>
