+++ src/actions/transforms/invert_action.cpp
@@ -4,8 +4,8 @@
 
 #include <fullscore/actions/transforms/invert_action.h>
 
-#include <fullscore/transforms/invert.h>
 #include <fullscore/models/note.h>
+#include <fullscore/transforms/invert_transform.h>
