+++ src/actions/transforms/octatonic_1_transform_action.cpp
@@ -3,7 +3,7 @@
 
 #include <fullscore/actions/transforms/octatonic_1_transform_action.h>
 
-#include <fullscore/transforms/octatonic_1.h>
+#include <fullscore/transforms/octatonic_1_transform.h>
 #include <fullscore/models/note.h>
