+++ src/actions/transforms/remove_dot_transform_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/transforms/remove_dot_transform_action.h>
 
-#include <fullscore/transforms/remove_dot.h>
+#include <fullscore/transforms/remove_dot_transform.h>
 #include <fullscore/models/note.h>
