+++ src/actions/transforms/transpose_up_action.cpp
@@ -4,8 +4,8 @@
 
 #include <fullscore/actions/transforms/transpose_up_action.h>
 
-#include <fullscore/transforms/transpose.h>
 #include <fullscore/models/note.h>
+#include <fullscore/transforms/transpose_transform.h>
