+++ src/transforms/erase_note_transform.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/transforms/erase_note.h>
+#include <fullscore/transforms/erase_note_transform.h>
 
 #include <algorithm>
