+++ tests/transforms/copy_test.cpp
@@ -3,7 +3,7 @@
 
 #include <gtest/gtest.h>
 
-#include <fullscore/transforms/copy.h>
+#include <fullscore/transforms/copy_transform.h>
 
 #include <fullscore/models/measure_grid.h>
