+++ tests/transforms/reference_test.cpp
@@ -3,7 +3,7 @@
 
 #include <gtest/gtest.h>
 
-#include <fullscore/transforms/reference.h>
+#include <fullscore/transforms/reference_transform.h>
 
 #include <fullscore/models/measure_grid.h>
